chat_queue_handler/requirements.txt
flask
requests
google-cloud-vision